Calendar: Start Month View on current week
Currently the Month View defaults to always start with the first
week of a month. Like this, one can see all the weeks of a
particular  month.

Some people are more interested in seeing upcoming events/weeks rather
than passed events/weeks.
There should be an option to let the Month View start with the current
week. Showing the weeks to come, up to a range of one month.
This would be equivalent to scrolling down to the current week when
opening the Month View while month-scroll-by-week is enabled.

This change
- adds an option in Edit->Preferences->Calendar and Tasks->Display
- the option is disabled by default
- enables above behavior in e-cal-shell-content.c
- since List View gets its date range the same way as Month View,
this also changes the behavior of List View to show respective ranges.

https://gitlab.gnome.org/GNOME/evolution/issues/734