Add an option to specify minimum font size for WebKitGTK
Evolution used to force some minimum font size in the code, but it
was removed to be able to properly show <small>, <xsmall> and such
texts in the web views. Unfortunately, it also means that if the sender
keeps such small font it can be hard to read to the recipient, thus
let's add the option, thus the user can set it up to his/her needs.