Add an option to show uncompleted tasks
Currently "Active Tasks" shows tasks due in the next 365
days. A common use-case is that one creates a task (e.g. on
a mobile device) and does not set an explicit due date.
Such tasks would not be listed under "Active Tasks". Thus
create a new filter "Uncompleted Tasks" that picks up
uncompleted tasks irrespective of the existence of a due
date.

Closes https://gitlab.gnome.org/GNOME/evolution/-/issues/858 .