ESourceSelector: Update GtkTreeStore values only if changed
The e_source_selector_set_source_connection_status() can be called
when the GtkTreeView updates itself, which can, even when the value
doesn't change, invoke an "accessible changed" signal, which then
can call the function again. Avoid the recursion by checking whether
the value really changed before setting it to the GtkTreeStore.
Do it also for the tooltip, to be consistent.