M!80 - help: Change instances of "Apply" to "OK"
The text on the button used to confirm, e.g., the creation of a new
calendar (in the "New Calendar" window) now reads "OK" instead of
"Apply", but the help pages for add/using various calendars still use
"Apply". The same applies to windows and help pages for adding/using
various address books and task/memo lists.

Change instances of "Apply" to "OK" for help pages that link to the
following help pages:

* calendar-using-several-calendars
* contacts-using-several-addressbooks
* tasks-using-several-tasklists

Closes https://gitlab.gnome.org/GNOME/evolution/-/merge_requests/80