Calendar: Not all task lists included in the task panel
This could happen when the Evolution had been started in the Tasks view
and then switched to the Calendar view, the task panel was left empty.
It was because all the task lists had been opened already, thus none
emitted the "client-opened" signal, thus the related data model did
not have added the clients.

This applies to the memos panel as well.