Use gtk_widget_is_visible() instead of gtk_widget_get_visible() on some places
The changed places rely on the visibility of the widget including
the parent widgets, not only whether the GtkWidget::visible is set
to true.

An example is when replying to a message when the preview panel is
turned off while it had been showing a different message in it
before being hid. In that case the composer opened with the message
in the hidden preview panel, instead of the message selected
in the message list.