EContactEditor: Use a GWeakRef on the editor when opening target client
The open of the target client can take long time in some cases. Use
a GWeakRef on the editor to properly detect when the editor is closed
while the client is opening (the editor itself could not be freed
meanwhile, but its widgets could be freed on the editor close).