MarkdownEditor: Change how cursor position is tracked in side-by-side preview
Inserting helper text in the markdown source can cause breakage of
the generated HTML, for example when the cursor is in the middle or
on a boundary of a markdown tag. The cmark supports writing source
position into the generated HTML, thus use that instead.

Also track cursor movement and update the preview scroll position,
if needed.