EMFolderProperties: Ignore persistent folder properties without 'blurb'
The blurb is supposed to be a localized text shown on the option.
Ignore properties, which do not have it set, instead of claiming
a runtime warning on the console. This allows Camel providers to
have stored custom properties on the folders, without reflecting
them in the folder Properties dialog.