Tweak how long to wait for a 'connected' state when opening an EClient
The 'connected' state is not reached when the data store is not available,
which means to wait for the whole set time. Use shorter times, also because
most of the backends support offline mode, thus the changes can be propagated
to the server later.

Related to https://gitlab.gnome.org/GNOME/evolution-data-server/-/issues/489