I#2450 - CompEditor: Enhance time re-calculation on overlap change
This covers the case when the start time is set after the end time or
when the end time is set before the start time. This will use the current
duration of the event for the shift of the other time, not the one
stored in the component, which is most likely outdated.

Closes https://gitlab.gnome.org/GNOME/evolution/-/issues/2450