EContactCard: Card sometimes hides values
The value grid height is used to determine how many lines of information
can be shown inside the card. When the grid had set the `valign` to `start`,
then once it was shrunk due to no lines of text shown in it, it could not
be made taller, because no labels fit into height of 1px. Making it `fill`
the assigned card height allows using consistent height for the values,
regardless which had been shown earlier and which not.