Composer: Ask before sending mail with an Autocrypt header
The option to send mail with the Autocrypt header is enabled by default
and the public PGP key is looked for by mail, when the user doesn't
have set a specific key ID, thus to avoid an unwanted send of the public
key do ask before doing so. The option to ask is also per-account.