I#2665 - Tasks: Add Status=Needs Action filter option
Adds the task type of 'Needs Action' in the Advanced Search feature in the Tasks
view. This task type can already be set in multiple ways, but now it can also be
searched for.

Related specification is RFC 2445 section 4.8.1.11 Status
https://www.ietf.org/rfc/rfc2445.txt

The implementation in the Evolution Data Server already exists.

Tested and confirmed working.

Closes https://gitlab.gnome.org/GNOME/evolution/-/issues/2665

Signed-off-by: Nico Rikken <nico@nicorikken.eu>