EAttachment: Check file_info contains display-name before accessing it
There could be shown a runtime warning about the attachment's file_info
not containing the display-name attribute while the code still tried
to access it, which is considered a bug in the code using glib2 API,
thus verify the attribute is available before accessing it. Also access
it only when needed, not always.