EWebView: Just loaded remote images may be left invisible
While calculating image size for the preview width, the images which are
being loaded report their width as 0px, which can be preserved due to it
can fix the preview width constraint, even though the image is 0px because
it's not loaded yet. The image does not enlarge to its expected size after
it's loaded due to this, effectively making the image invisible.