Guard most of the g_file_info_get_...() calls with checks for attribute existence
The g_file_info_get_...() functions can claim a runtime warning on the terminal
when the associated attribute is not present in the file info. As not on every
place can be assured the requested attribute is present, add the guards to only
read the attribute when it exists in the file info structure.