I#2803 - Mail: Add a 'Correspondents' column to the message list
Check whether the From address is one of the enabled configured accounts and
if so, then consider the message as being sent, not received; this fails when
sending from one configured account to another configured account, thus first
guess no received date is a sent mail, which does not work for some providers,
like Maildir, which sets the received date from the filename. Trying address
of the account of the selected folder won't work for On This Computer folders.
In other words, this is a pure guess, which has many counter cases.

Closes https://gitlab.gnome.org/GNOME/evolution/-/issues/2803