I#2941 - mail-autoconfig: Use real user e-mail in the request
Some services require real user email to provide accurate account
configuration information for that user. The MD5 email hash is
a specific thing in the Evolution, nothing most of the service
providers know about or may have an intention to implement.
Because there are other ways how to steal the user mail,
and Thunderbird does this from the beginning, then finally
stop sending the fake email address and follow the lead.

The MD5 argument is left there, because there's at least one service
using it (it was added for them).

Closes https://gitlab.gnome.org/GNOME/evolution/-/issues/2941