Switcher button icon does not change when action's icon name changes
The Mail switcher button changes its icon to an "unread mail", when
a new mail is received and then back to a "mail", when any mail is read.
This did not work properly, because the switcher button's properties
were not updated when the action's property changed.

Cover also the label property, just in case.