Preview layout setting does not stick after app restart
The change from classic/horizontal to vertical is saved into the GSettings,
but on start the state was restored to the second action of this radio group,
while the group was not set yet, due to the EUIManager being frozen at that
time, thus when the EUIMnager had been unfrozen, the value of the first
action in the group was used to update the group value, which did not reflect
the value read from the GSettings.

Instead of shifting the binding settings to a place where the EUIManager is truly
unfrozen, simply use the first action in the group for the GSettings key.