EWeekView: Scroll by shown weeks when scrolling by month
This covers for example a mouse wheel scroll, when "scroll by month" is
enabled by the user.

The 4 weeks scroll does not work for a case when less than 4 weeks are shown,
because it skips those in-between weeks. When more than 4 weeks are shown
the view repeats the extra weeks, which can be also unexpected.