mail-vfolder: Remove unneeded runtime warning
It could show up when a vFolder was renamed, but as the vStore itself
renamed it, the old name is changed already in the internal structures,
thus the warning is false. The rename notification is received asynchronously,
thus it cannot be simply blocked while calling camel_store_rename_folder_sync().